backend/requirements.txt
# KdB Assistant Backend

fastapi==0.115.6
uvicorn[standard]==0.34.0
pandas==2.2.3
openpyxl==3.1.5
httpx==0.28.1
beautifulsoup4==4.12.3
chromadb==0.5.23
google-generativeai==0.8.3
python-dotenv==1.0.1
pydantic==2.10.3
sse-starlette==2.2.1
python-multipart==0.0.18
